Added SyncThing
Copied from Runtipi Official AppStore, but updated removed ROOT_FOLDER_HOST mapping, as planning to use symlinks in data directory for folders to sync. Changed description for SYNCTHING_HOSTNAME variable, as this is not a HOSTNAME, but the DEVICE NAME that SyncThing will use. While this can be changed via the UI later, can just set it on creation too.